feat(api): implement from_json method for datahub_actions.event.event_envelope.EventEnvelope
