+++ datahub-actions/src/datahub_actions/action/action.py
@@ -16,7 +16,7 @@ from abc import ABCMeta, abstractmethod
 
 from datahub.ingestion.api.closeable import Closeable
 
-from datahub_actions.event.event import EventEnvelope
+from datahub_actions.event.event_envelope import EventEnvelope
 from datahub_actions.pipeline.pipeline_context import PipelineContext
