+++ datahub-actions/src/datahub_actions/source/event_source.py
@@ -17,7 +17,7 @@ from typing import Iterable
 
 from datahub.ingestion.api.closeable import Closeable
 
-from datahub_actions.event.event import EventEnvelope
+from datahub_actions.event.event_envelope import EventEnvelope
 from datahub_actions.pipeline.pipeline_context import PipelineContext
